# ARROW-10772: [Rust] Speed up take by writing to buffer