@ritchie46 2020-11-30
Already experimented a bit for the stringarray variant. https://github.com/ritchie46/polars/pull/193